# server/resource/Manage_Film_Example_db/Film.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE Film.js PLEASE EDIT ../custom/FilmCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Manage_Film_Example_db = require('../../db/Manage_Film_Example_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/FilmCustom.js');

/*
 * SCHEMA DB Film
 * 
	{
		genre: {
			type: 'String',
			enum : ["Action","Crime","Fantasy","Horror"]
		},
		title: {
			type: 'String', 
			required : true
		},
		year: {
			type: 'Number'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		cast: [{
			type: Schema.ObjectId,
			ref : "Film"
		}],
		filmMaker: {
			type: Schema.ObjectId,
			required : true,
			ref : "Film"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * FilmService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/films', function (req, res) {
	obj = new db_Manage_Film_Example_db.Film(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * FilmService.delete
 *   @description CRUD ACTION delete
 *
 */
app['delete'](properties.api + '/films/:id', function (req, res) {
	db_Manage_Film_Example_db.Film.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * FilmService.findBycast
 *   @description CRUD ACTION findBycast
 *
 */
app.get(properties.api + '/films/findBycast/:key', function (req, res) {

	db_Manage_Film_Example_db.Film.find({ 'cast' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * FilmService.findByfilmMaker
 *   @description CRUD ACTION findByfilmMaker
 *
 */
app.get(properties.api + '/films/findByfilmMaker/:key', function (req, res) {

	db_Manage_Film_Example_db.Film.find({ 'filmMaker' : req.params.key}).exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
	
});

/**
 * FilmService.get
 *   @description CRUD ACTION get
 *
 */
app.get(properties.api + '/films/:id', function (req, res) {
	db_Manage_Film_Example_db.Film.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * FilmService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/films', function (req, res) {
	db_Manage_Film_Example_db.Film.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * FilmService.update
 *   @description CRUD ACTION update
 *
 */
app.post(properties.api + '/films/:id', function (req, res) {
	db_Manage_Film_Example_db.Film.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */