# server/resource/Manage_Film_Example_db/FilmMaker.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE FilmMaker.js PLEASE EDIT ../custom/FilmMakerCustom.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
const app = require('../../app.js');
const db_Manage_Film_Example_db = require('../../db/Manage_Film_Example_db_schema.js')
const properties = require('../../properties.js');
const handleError = require('../../security/util.js').handleError;
require('./custom/FilmMakerCustom.js');

/*
 * SCHEMA DB FilmMaker
 * 
	{
		name: {
			type: 'String', 
			required : true
		},
		surname: {
			type: 'String'
		},
		//RELATIONS
		
		
		//EXTERNAL RELATIONS
		
		filmMaker: {
			type: Schema.ObjectId,
			required : true,
			ref : "Film"
		},
		
	}
 * 
 */



//CRUD METHODS


/**
 * FilmMakerService.create
 *   @description CRUD ACTION create
 *
 */
app.post(properties.api + '/filmmakers', function (req, res) {
	obj = new db_Manage_Film_Example_db.FilmMaker(req.body);
	obj.save(function (err) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * FilmMakerService.delete
 *   @description CRUD ACTION delete
 *
 */
app['delete'](properties.api + '/filmmakers/:id', function (req, res) {
	db_Manage_Film_Example_db.FilmMaker.findByIdAndRemove(req.params.id, function (err) {
		  if (err) return handleError(err, res);
		  res.send(err);
	});
});

/**
 * FilmMakerService.get
 *   @description CRUD ACTION get
 *
 */
app.get(properties.api + '/filmmakers/:id', function (req, res) {
	db_Manage_Film_Example_db.FilmMaker.findOne({_id:req.params.id}).exec(function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});

/**
 * FilmMakerService.list
 *   @description CRUD ACTION list
 *
 */
app.get(properties.api + '/filmmakers', function (req, res) {
	db_Manage_Film_Example_db.FilmMaker.find().exec(function (err, list) {
		if (err) return handleError(err, res);
		res.send(list);
	});
});

/**
 * FilmMakerService.update
 *   @description CRUD ACTION update
 *
 */
app.post(properties.api + '/filmmakers/:id', function (req, res) {
	db_Manage_Film_Example_db.FilmMaker.findByIdAndUpdate(req.params.id, req.body, {'new': true}, function (err, obj) {
		if (err) return handleError(err, res);
		res.send(obj);
	});
});


/*
 * CUSTOM SERVICES
 * 
 *	These services will be overwritten and implemented in  Custom.js
 */